feat: Added ErrorBoundary component
