--- app/admin/not-found.tsx
@@ -1,11 +1,0 @@
-import Link from "next/link";
-
-export default function NotFound() {
-  return (
-    <div>
-      <h2 className="text-3xl">Page not found</h2>
-      <p>Could not find requested resource</p>
-      <Link href="/admin">&larr; Home</Link>
-    </div>
-  );
-}
